Refactor GeneticAlgorithm and update related classes
Refactored the GeneticAlgorithm class with a new constructor, added fields, and methods for execution and population initialization. Updated ProductsExtension to cast dimensions to float. Enhanced Chromosome class with new namespace, using directives, and constructor changes. Introduced ProductPlacement class. Reformatted cubeB initialization in SAT3D.cs. Added FitnessCalculator class in Services namespace.

+++ 3D Bin Packing Problem/Extensions/ProductsExtension.cs
@@ -8,14 +8,14 @@ namespace _3D_Bin_Packing_Problem.Extensions
         public static Vector3[] ToVector3(this Product product, Vector3 middle)
         {
             Vector3[] poly = new Vector3[8];
-            poly[0] = new Vector3(middle.X - product.Length / 2, middle.Y - product.Width / 2, middle.Z - product.Height / 2);
-            poly[1] = new Vector3(middle.X + product.Length / 2, middle.Y - product.Width / 2, middle.Z - product.Height / 2);
-            poly[2] = new Vector3(middle.X + product.Length / 2, middle.Y + product.Width / 2, middle.Z - product.Height / 2);
-            poly[3] = new Vector3(middle.X - product.Length / 2, middle.Y + product.Width / 2, middle.Z - product.Height / 2);
-            poly[4] = new Vector3(middle.X - product.Length / 2, middle.Y - product.Width / 2, middle.Z + product.Height / 2);
-            poly[5] = new Vector3(middle.X + product.Length / 2, middle.Y - product.Width / 2, middle.Z + product.Height / 2);
-            poly[6] = new Vector3(middle.X + product.Length / 2, middle.Y + product.Width / 2, middle.Z + product.Height / 2);
-            poly[7] = new Vector3(middle.X - product.Length / 2, middle.Y + product.Width / 2, middle.Z + product.Height / 2);
+            poly[0] = new Vector3(middle.X - (float)product.Length / 2, middle.Y - (float)product.Width / 2, middle.Z - (float)product.Height / 2);
+            poly[1] = new Vector3(middle.X + (float)product.Length / 2, middle.Y - (float)product.Width / 2, middle.Z - (float)product.Height / 2);
+            poly[2] = new Vector3(middle.X + (float)product.Length / 2, middle.Y + (float)product.Width / 2, middle.Z - (float)product.Height / 2);
+            poly[3] = new Vector3(middle.X - (float)product.Length / 2, middle.Y + (float)product.Width / 2, middle.Z - (float)product.Height / 2);
+            poly[4] = new Vector3(middle.X - (float)product.Length / 2, middle.Y - (float)product.Width / 2, middle.Z + (float)product.Height / 2);
+            poly[5] = new Vector3(middle.X + (float)product.Length / 2, middle.Y - (float)product.Width / 2, middle.Z + (float)product.Height / 2);
+            poly[6] = new Vector3(middle.X + (float)product.Length / 2, middle.Y + (float)product.Width / 2, middle.Z + (float)product.Height / 2);
+            poly[7] = new Vector3(middle.X - (float)product.Length / 2, middle.Y + (float)product.Width / 2, middle.Z + (float)product.Height / 2);
             return poly;
         }
     }
